Changed error message for when email is incorrect
Previously the error message for trying to log into a non-existent
account was different from the one for trying to log into an existing
account, but with the wrong password. This difference weakened security
because an attacker could figure out which e-mails were actually in the
database, and I solved this issue in this commit.

I also got rid of an extraneous comment.